EN update.
Added english output option of vypoctinejistotu(), which was renamed to calcunc() but the vypoctinejistotu was aliased to it for compatibility. Similarly the zaoknem() was renamed and aliased back.

The calcunc() has now the option to choose standard deviation type (biased/unbiased/of mean). Default remains the std. dev. of mean.